[CHORE] Request와 Response를 분리하였습니다.
