+++ app/src/main/java/org/sopt/dosopttemplate/network/dto/req/SignInRequest.kt
@@ -1,13 +1,13 @@
-package org.sopt.dosopttemplate.network.dto
+package org.sopt.dosopttemplate.network.dto.req
 
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class UserInfoResponse(
-    @SerialName("nickname")
-    val nickname: String = "",
+data class SignInRequest(
+    @SerialName("password")
+    val password: String = "",
     @SerialName("username")
     val username: String = ""
 )
