+++ app/src/main/java/org/sopt/dosopttemplate/network/dto/req/SignUpRequest.kt
@@ -1,11 +1,13 @@
-package org.sopt.dosopttemplate.network.dto
+package org.sopt.dosopttemplate.network.dto.req
 
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class SignInRequest(
+data class SignUpRequest(
+    @SerialName("nickname")
+    val nickname: String = "",
     @SerialName("password")
     val password: String = "",
     @SerialName("username")
