+++ app/src/main/java/org/sopt/dosopttemplate/network/dto/res/FriendListResponse.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.network.dto
+package org.sopt.dosopttemplate.network.dto.res
 
 
 import kotlinx.serialization.SerialName
