+++ app/src/main/java/org/sopt/dosopttemplate/network/dto/res/UserInfoResponse.kt
@@ -1,13 +1,11 @@
-package org.sopt.dosopttemplate.network.dto
+package org.sopt.dosopttemplate.network.dto.res
 
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class SignInResponse(
-    @SerialName("id")
-    val id: Int = 0,
+data class UserInfoResponse(
     @SerialName("nickname")
     val nickname: String = "",
     @SerialName("username")
